requirements.txt
Django==5.2.12
django-unfold==0.88.0
python-dotenv==1.2.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
psycopg2-binary==2.9.11
django-filter==25.2
django-cors-headers==4.9.0